+++ cogs/help.py
@@ -1,6 +1,7 @@
 import discord
 from discord.ext import commands
 
+from db.dbase import DBase
 from cogs.utils import constants
 from cogs.utils.messages import MessageManager
 
@@ -17,13 +18,17 @@ class Help:
         user = ctx.message.author
         manager = MessageManager(self.bot, user, ctx.message.channel, ctx.message)
 
+        prefix = ""
+        with DBase() as db:
+            prefix = db.get_prefix(ctx.message.server.id)
+
+        text = ("**{0}event create** - create a new event\n"
+              + "**{0}event delete <id>** - delete event with the given ID\n"
+              + "**{0}role <class>** - choose which role you intend on playing in D2\n"
+              + "**{0}roster** - display the selected role of all members\n"
+              + "**{0}prefix** - change the command prefix").format(prefix)
         help = discord.Embed(title="Available Commands", color=constants.BLUE)
-        help.add_field(name="Events",
-                       value="**!event create** - create a new event\n"
-                           + "**!event delete <id>** - delete event with the given ID")
-        help.add_field(name="Roster",
-                       value="**!role <class>** - lets others know class you intend on playing in D2\n"
-                           + "**!roster** - display a list of what classes everyone intends on playing in D2")
+        help.description = text
 
         await manager.say(help, embed=True, delete=False)
         await manager.clear()